requirements.txt
geopandas==0.14.2
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
tqdm==4.65.0
